requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic>=2.7.0
requests>=2.32.3
PyJWT>=2.9.0
python-dotenv>=1.0.1
# CLI wrappers rely on external binaries: semgrep, gitleaks, syft, grype (installed in Docker image)


